+++ niffler-userdata/src/test/java/guru/qa/niffler/controller/v2/FriendsV2ControllerTest.java
@@ -1,4 +1,4 @@
-package guru.qa.niffler.controller;
+package guru.qa.niffler.controller.v2;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
